Fix zoom with center fixed
In Camera class changed offsetX and offsetY from type int to
type double. Before they represented the offset in pixels now
they represent the offset in fraction of tile size. This avoids
loss due to integer division which allows implementation of zoom
with the center of the visible grid being fixed on screen.